Implmented Heap based solution
The idea here is to maintain a min heap and a max heap
with a even number of items in each, where all of the values
larger than the median are in the min heap and all the values
smaller than the median are in the max heap.

i.e. the median is the first value of the heap tha is larger by
one element, or the average of the two heads if they have
the same count.

When a new item is added we compare it to the current median
to decide what heap it should go in.
If this makes the heaps unballanced we pop from one heap and
push to the other.

----

Future cleanup would be to implment a clean generic Heap<T>
that can be constructed with a comparer allowing it to be a min
or a max heap. (Or just use one form a library)